* Fixed wrong timer handling while other player controlled a player's turn. Attackers now marked with blue frame. Playable cards have a violet frame.  If a player has to select cards from hand, the possible cards are marked yellow now. Discard of multiple cards now marks already selected cards and happens in one selection.

+++ Mage/src/mage/game/GameImpl.java
@@ -988,7 +988,7 @@ public abstract class GameImpl implements Game, Serializable {
         int deduction = 1;
         if (freeMulligans > 0) {
             if (usedFreeMulligans != null && usedFreeMulligans.containsKey(player.getId())) {
-                int used = usedFreeMulligans.get(player.getId()).intValue();
+                int used = usedFreeMulligans.get(player.getId());
                 if (used < freeMulligans ) {
                     deduction = 0;
                 }
